requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.1
Django==3.2.5
django-crispy-forms==1.12.0
django-ratelimit==2.0.0
idna==3.2
itsdangerous==2.0.1
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-cache==0.7.1
six==1.16.0
sqlparse==0.4.1
url-normalize==1.4.3
urllib3==1.26.6
